requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
backports.ssl_match_hostname==3.7.0.1
cfgv==3.4.0
decorator==5.2.1
distlib==0.4.0
Django==5.1.4
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
executing==2.2.0
filelock==3.18.0
identify==2.6.12
inflection==0.5.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
matplotlib-inline==0.1.7
nodeenv==1.9.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
pre_commit==4.0.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
ruff==0.8.4
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
uritemplate==4.2.0
virtualenv==20.33.0
wcwidth==0.2.13
